[feat] 조직 등록 및 인사발령 코드 리팩토링
