+++ DDIS_HR/src/main/java/com/ddis/ddis_hr/organization/command/application/dto/DepartmentRequestDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class DepartmentRequestDTO {
 
     private String departmentName;
-    private String headName;
+    private Long headId;
 }
