+++ DDIS_HR/src/main/java/com/ddis/ddis_hr/organization/query/dao/IntroductionMapper.java
@@ -35,4 +35,6 @@ public interface IntroductionMapper {
     List<PositionDTO> selectPositionsByJobId(Long jobId);
 
     List<RankDTO> selectRanksByJobId(Long jobId);
+
+    List<EmployeeQueryDTO> findAllEmployees();
 }
